+++ components/layout/Layout.tsx
@@ -31,32 +31,33 @@ const Layout: FunctionComponent<LayoutProps> = ({
   const siteUrl = process.env.NEXT_PUBLIC_SITE_URL || "";
   const twitterImageContentUrl = siteImage ? siteImage : SITE_IMAGE;
   const twitterImageContent = `${siteUrl}${twitterImageContentUrl}`;
+  const pageTitle = title
+    ? `${uppercaseFirst(title)} - ${SITE_NAME}`
+    : SITE_NAME;
+  const pageDescription = metaDescription
+    ? metaDescription
+    : `${SITE_NAME} - ${SITE_DESCRIPTION}`;
+  const propertiesTitle = title ? uppercaseFirst(title) : SITE_NAME;
+  const propertiesDescription = metaDescription
+    ? metaDescription
+    : SITE_DESCRIPTION;
 
   return (
     <div className={styles.container}>
       <Head>
-        <title>
-          {title ? `${uppercaseFirst(title)} - ${SITE_NAME}` : SITE_NAME}
-        </title>
-        <meta
-          name="description"
-          content={
-            metaDescription
-              ? metaDescription
-              : `${SITE_NAME} - ${SITE_DESCRIPTION}`
-          }
-        />
+        <title>{pageTitle}</title>
+        <meta name="description" content={pageDescription} />
 
         {/* Google Properties */}
-        <meta itemProp="name" content={SITE_NAME} />
-        <meta itemProp="description" content={SITE_DESCRIPTION} />
+        <meta itemProp="name" content={propertiesTitle} />
+        <meta itemProp="description" content={propertiesDescription} />
         <meta itemProp="image" content={siteImage ? siteImage : SITE_IMAGE} />
 
         {/* Open Graph Properties */}
         <meta property="og:url" content="https://www.devoreur2code.com" />
         <meta property="og:type" content="website" />
-        <meta property="og:title" content={SITE_NAME} />
-        <meta property="og:description" content={SITE_DESCRIPTION} />
+        <meta property="og:title" content={propertiesTitle} />
+        <meta property="og:description" content={propertiesDescription} />
         <meta
           property="og:image"
           content={siteImage ? siteImage : SITE_IMAGE}
@@ -65,8 +66,8 @@ const Layout: FunctionComponent<LayoutProps> = ({
         {/* Twitter Properties */}
         <meta name="twitter:card" content="summary" />
         <meta name="twitter:site" content="@devoreur2code" />
-        <meta name="twitter:title" content={SITE_NAME} />
-        <meta name="twitter:description" content={SITE_DESCRIPTION} />
+        <meta name="twitter:title" content={propertiesTitle} />
+        <meta name="twitter:description" content={propertiesDescription} />
         <meta name="twitter:image" content={twitterImageContent} />
 
         <link rel="icon" href={FAVICON_URL} />