+++ pages/blog/[slug].tsx
@@ -66,6 +66,7 @@ export async function getStaticProps({ params }: Params) {
     "coverImage",
     "ogImage",
     "tags",
+    "excerpt",
   ]);
   const content = await markdownToHtml(post.content || "");